requirements.txt
django== 5.2.8
djangorestframework== 3.16.1